Refactor OSTree tests and utils
Create the ostree_utils module which will contain any utilities specific
to OSTree. Create the test_crud module and move the existing crud tests
into it.
